Builds ACR, GCR credential providers in CI (#51191)
This change builds the GCR and ACR credential providers in CI.

This is due to bug OCPBUGS-30970. Initially, we did not think we needed
builds in 4.15, but due to RHEL workers not always being upgraded before
the cluster, we do (see bug).

This builds the RPMs in CI to ensure no 4.15 backports break the RPM
builds.