HIVE-2400: Drop hive's build root for ocp golang (#53198)
The alternative of bumping hive repo's build/build-images/Dockerfile to
rhel9 is not immediately useful due to the fact that epel 9 does not
carry tcping. It was only installed for some 2018 troubleshooting of the
build stages. It's probably safe to drop this in favor of a more common
build root.

Signed-off-by: Antoni Segura Puimedon <antoni@redhat.com>