medik8s: fix tide and cleanup (#53781)
* medik8s: fix tide again and cleanup

- add missing 4.16, and all 4.17 configs
- remove configs of unsupported versions
- move tide config to org level for deduplication
- make jobs optional; they will be made required by branch protection

Reason for the latter:
Our tests use skip_if_only_changed, which makes them not required,
what we don't want. That's why we added branch protection config.
However, for tide those tests are configured twice now: as "required"
from branch protection, and "required if present" from job config. That failed
tide. To fix this, we make the jobs optional. Letting tide ignore branch protection
would be nicer, but it's not possible to override that setting.

Signed-off-by: Marc Sluiter <msluiter@redhat.com>

* update jobs

Signed-off-by: Marc Sluiter <msluiter@redhat.com>

---------

Signed-off-by: Marc Sluiter <msluiter@redhat.com>