WINC-1310: Remove redundant AWS upgrade job and use Windows Server 2019 in Azure operator test (#54158)
* [winc] Remove redundant AWS upgrade job

Removes the old upgrade test, as it has been replaced by the newer
upgrade test job for a while now.

Ran `make update`

* Use Windows Server 2019 in Azure test job

In order to maintain test coverage of Windows Server 2019 (supported until EOL),
the Azure operator test is changed to use 2019 instead of 2022.
The Azure upgrade test still uses 2022.
The AWS upgrade job used to test with 2019 nodes but that job is now removed.