Update ROSA cluster version for compliance operator CI (#54704)
* Use ipi-install-rbac for Compliance Operator e2e ROSA tests

The ROSA HCP workflow needs access to the build cluster registry to pull
images. This commit updates the test config so that it has access and is
able to stand up a ROSA HCP cluster for CI.

* Update ROSA cluster version for compliance operator CI

We were relying on OpenShift version 4.15.8, which isn't available
through ROSA HCP anymore. Let's use a newer, supported version so we
don't get the following error:

   Requested cluster version not available!